python-project/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
homebrew-pypi-poet
